fixed 6 points of 10
fixed field capacity pin error. now if center is pressed more than 3
seconds enter in selection mode, erased % on HSO row. fixed some clear
LCD bugs. finished about screen.

+++ SAC_prueba2/Relay.h
@@ -42,9 +42,9 @@ typedef struct _Relay Relay;
 
   
   struct _Relay {
-  int gpio_pin;
-  int role;
-  int state;
+  byte gpio_pin;
+  byte role;
+  byte state;
  };
